requirements.txt
python-dotenv
crewai
streamlit